refactor: rename rv32_alu module to be more general + small PR comments
