+++ vm/src/arch/chips.rs
@@ -28,8 +28,8 @@ use crate::{
     memory::MemoryChipRef,
     modular_addsub::ModularAddSubChip,
     modular_multdiv::ModularMultDivChip,
+    new_alu::Rv32ArithmeticLogicChip,
     program::{ExecutionError, Instruction, ProgramChip},
-    rv32_alu::Rv32ArithmeticLogicChip,
     shift::ShiftChip,
     ui::UiChip,
     uint_multiplication::UintMultiplicationChip,
