+++ vm/src/lib.rs
@@ -9,8 +9,8 @@ pub mod hashes;
 pub mod memory;
 pub mod modular_addsub;
 pub mod modular_multdiv;
+pub mod new_alu;
 pub mod program;
-pub mod rv32_alu;
 /// SDK functions for running and proving programs in the VM.
 #[cfg(feature = "sdk")]
 pub mod sdk;
