+++ vm/src/new_alu/tests.rs
@@ -1,5 +1,5 @@
 use super::integration::solve_alu;
-use crate::arch::instructions::RV32AluOpcode;
+use crate::arch::instructions::AluOpcode;
 
 const RV32_NUM_LIMBS: usize = 4;
 const RV32_LIMB_BITS: usize = 8;
@@ -9,7 +9,7 @@ fn solve_add_sanity_test() {
     let x: [u32; RV32_NUM_LIMBS] = [229, 33, 29, 111];
     let y: [u32; RV32_NUM_LIMBS] = [50, 171, 44, 194];
     let z: [u32; RV32_NUM_LIMBS] = [23, 205, 73, 49];
-    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(RV32AluOpcode::ADD, &x, &y);
+    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(AluOpcode::ADD, &x, &y);
     for i in 0..RV32_NUM_LIMBS {
         assert_eq!(z[i], result[i])
     }
@@ -20,7 +20,7 @@ fn solve_sub_sanity_test() {
     let x: [u32; RV32_NUM_LIMBS] = [229, 33, 29, 111];
     let y: [u32; RV32_NUM_LIMBS] = [50, 171, 44, 194];
     let z: [u32; RV32_NUM_LIMBS] = [179, 118, 240, 172];
-    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(RV32AluOpcode::SUB, &x, &y);
+    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(AluOpcode::SUB, &x, &y);
     for i in 0..RV32_NUM_LIMBS {
         assert_eq!(z[i], result[i])
     }
@@ -31,7 +31,7 @@ fn solve_xor_sanity_test() {
     let x: [u32; RV32_NUM_LIMBS] = [229, 33, 29, 111];
     let y: [u32; RV32_NUM_LIMBS] = [50, 171, 44, 194];
     let z: [u32; RV32_NUM_LIMBS] = [215, 138, 49, 173];
-    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(RV32AluOpcode::XOR, &x, &y);
+    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(AluOpcode::XOR, &x, &y);
     for i in 0..RV32_NUM_LIMBS {
         assert_eq!(z[i], result[i])
     }
@@ -42,7 +42,7 @@ fn solve_or_sanity_test() {
     let x: [u32; RV32_NUM_LIMBS] = [229, 33, 29, 111];
     let y: [u32; RV32_NUM_LIMBS] = [50, 171, 44, 194];
     let z: [u32; RV32_NUM_LIMBS] = [247, 171, 61, 239];
-    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(RV32AluOpcode::OR, &x, &y);
+    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(AluOpcode::OR, &x, &y);
     for i in 0..RV32_NUM_LIMBS {
         assert_eq!(z[i], result[i])
     }
@@ -53,7 +53,7 @@ fn solve_and_sanity_test() {
     let x: [u32; RV32_NUM_LIMBS] = [229, 33, 29, 111];
     let y: [u32; RV32_NUM_LIMBS] = [50, 171, 44, 194];
     let z: [u32; RV32_NUM_LIMBS] = [32, 33, 12, 66];
-    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(RV32AluOpcode::AND, &x, &y);
+    let result = solve_alu::<RV32_NUM_LIMBS, RV32_LIMB_BITS>(AluOpcode::AND, &x, &y);
     for i in 0..RV32_NUM_LIMBS {
         assert_eq!(z[i], result[i])
     }
