+++ vm/src/vm/segment.rs
@@ -44,8 +44,8 @@ use crate::{
     memory::{offline_checker::MemoryBus, MemoryChip, MemoryChipRef},
     modular_addsub::{ModularAddSubChip, SECP256K1_COORD_PRIME, SECP256K1_SCALAR_PRIME},
     modular_multdiv::ModularMultDivChip,
+    new_alu::{ArithmeticLogicIntegration, Rv32ArithmeticLogicChip},
     program::{bridge::ProgramBus, DebugInfo, ExecutionError, Program, ProgramChip},
-    rv32_alu::{ArithmeticLogicIntegration, Rv32ArithmeticLogicChip},
     shift::ShiftChip,
     ui::UiChip,
     uint_multiplication::UintMultiplicationChip,
